Exception and boolean return to CRUD methods

+++ src/main/java/br/bodoque/PrevalentObject.java
@@ -5,28 +5,42 @@ public abstract class PrevalentObject<T extends Prevalent> implements Prevalent 
 	private T prevalentObject;
 	private Long oID;
 
-	public void save() {
+	public boolean save() {
 		prevalentObject = whoAmI();
 		
-		//OMG not possible use is* in method names because FlexJson serialize it \o/
-		if (this.ehPrevalentObjectNotPersistedYet())
-			this.generateOIDForThisPrevalentObject();
+		if (isPrevalentObjectNotPersistedYet())
+			generateOIDForThisPrevalentObject();
 		
 		addPrevalentObjectToRepository(prevalentObject);
 		Command serializeCommand = createSerializeCommand(prevalentObject);
 		addCommandToLogList(serializeCommand);
+		return true;
 	}
 	
-	public void delete(boolean... shouldRaiseException) {
+	public boolean delete(boolean... shouldRaiseException) {
 		prevalentObject = whoAmI();
-		if (shouldRaiseException.length > 0) {
-			if(this.oID == null) { 
-				String message = String.format("Cannot remove object: %s. Only allowed to remove objects saved.", prevalentObject);
-				throw new CannotRemovePrevalentObjectException(message); 
-			} 
+		if (!raiseOrNotThisIsTheQuestion(this.getOID(), prevalentObject.getClass(), 
+										 shouldRaiseException)) {
+			return false;
 		}
 		deletePrevalentObject(prevalentObject, this.oID);
 		removeCommandFromLogList(prevalentObject);
+		return true;
+	}
+
+	private static boolean raiseOrNotThisIsTheQuestion(Long oID, Class<? extends Prevalent> 
+													prevalentObjectClass, boolean... shouldRaiseException) {
+		if(oID == null) { 
+			if (shouldRaiseException.length > 0 && shouldRaiseException[0]) {
+				System.out.println(1);
+				String message = String.format("Cannot remove object of class %s. " +
+						"Only allowed to remove objects saved.", prevalentObjectClass.getSimpleName());
+				throw new CannotDeletePrevalentObjectException(message); 
+			} else {
+				return false;
+			}
+		}
+		return true;
 	}
 
 	@SuppressWarnings("unchecked")
@@ -36,24 +50,26 @@ public abstract class PrevalentObject<T extends Prevalent> implements Prevalent 
 	}
 	
 	@SuppressWarnings("unchecked")
-	public static <T extends Prevalent> void save(T prevalentObject) {
+	public static <T extends Prevalent> boolean save(T prevalentObject) {
 		if (((PrevalentObject<T>) prevalentObject).getOID() == null)
 			((PrevalentObject<T>) prevalentObject).oID = Sequence.getNextOIDFor(prevalentObject.getClass());
 		
 		addPrevalentObjectToRepository(prevalentObject);
 		Command serializeCommand = createSerializeCommand(prevalentObject);
 		addCommandToLogList(serializeCommand);
+		return true;
 	}
 	
 	@SuppressWarnings("unchecked")
-	public static <T extends Prevalent> void delete(T prevalentObject) {
-		Long oId = ((PrevalentObject<T>) prevalentObject).getOID();
-		if (oId == null) { return ; }
-		deletePrevalentObject(prevalentObject, oId);
+	public static <T extends Prevalent> boolean delete(T prevalentObject, boolean... shouldRaiseException) {
+		Long oID = ((PrevalentObject<T>) prevalentObject).getOID();
+		if (!raiseOrNotThisIsTheQuestion(oID, prevalentObject.getClass(), shouldRaiseException)) {
+			return false;
+		}
+		deletePrevalentObject(prevalentObject, oID);
 		removeCommandFromLogList(prevalentObject);
+		return true;
 	}
-	
-
 
 	private static <T extends Prevalent> void removeCommandFromLogList(T prevalentObject) {
 		CommandLogList.removeCommand(new SerializeCommand<Prevalent>(prevalentObject));
@@ -76,7 +92,7 @@ public abstract class PrevalentObject<T extends Prevalent> implements Prevalent 
 		this.oID = Sequence.getNextOIDFor(prevalentObject.getClass());
 	}
 
-	private boolean ehPrevalentObjectNotPersistedYet() {
+	private boolean isPrevalentObjectNotPersistedYet() {
 		return this.oID == null;
 	}
